handle_connection を分離した
